+++ apps/posts/services/comment/comment_delete_services.py
@@ -1,0 +1,13 @@
+from typing import Any
+
+from apps.posts.exceptions.comment_exceptions import CommentForbiddenException
+from apps.posts.models.post_comment import PostComment
+
+
+def delete_comment(user: Any, comment: PostComment) -> None:
+    """
+    댓글 삭제 (작성자만 가능)
+    """
+    if comment.author != user:
+        raise CommentForbiddenException()
+    comment.delete()
