+++ apps/posts/views/__init__.py
@@ -1,9 +1,11 @@
-from .comment_views import (
-    PostCommentListCreateAPIView,
-    PostCommentRetrieveUpdateDestroyAPIView,
-)
+from .comment.comment_create_views import PostCommentCreateAPIView
+from .comment.comment_delete_views import PostCommentDeleteAPIView
+from .comment.comment_list_views import PostCommentListAPIView
+from .comment.comment_update_views import PostCommentUpdateAPIView
 
 __all__ = [
-    "PostCommentListCreateAPIView",
-    "PostCommentRetrieveUpdateDestroyAPIView",
+    "PostCommentListAPIView",
+    "PostCommentCreateAPIView",
+    "PostCommentUpdateAPIView",
+    "PostCommentDeleteAPIView",
 ]
